name refactor for mission nodes
